requirements.txt
backcall==0.2.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
fonttools==4.28.5
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.5
opencv-python==4.5.5.62
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.5.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pyzmq==22.3.0
six==1.16.0
tenacity==8.0.1
torch==1.10.1
torchsummary==1.5.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
wcwidth==0.2.5
